+++ components/detail-page/average-price-modal.tsx
