+++ components/detail-page/charts-section.tsx
