+++ components/detail-page/search-bar.tsx
