+++ components/detail-page/sentiment-score.tsx
