+++ components/detail-page/stock-info.tsx
