+++ components/detail-page/value-chain.tsx
